+++ internal/vm/bytecode/opcode.go
