+++ internal/vm/vmmem/memory.go
