+++ internal/vmcompiler/parser.go
@@ -5,11 +5,11 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/leonardinius/goloxvm/internal/bytecode"
-	"github.com/leonardinius/goloxvm/internal/scanner"
-	"github.com/leonardinius/goloxvm/internal/tokens"
-	"github.com/leonardinius/goloxvm/internal/vmobject"
-	"github.com/leonardinius/goloxvm/internal/vmvalue"
+	"github.com/leonardinius/goloxvm/internal/vm/bytecode"
+	"github.com/leonardinius/goloxvm/internal/vm/vmobject"
+	"github.com/leonardinius/goloxvm/internal/vm/vmvalue"
+	"github.com/leonardinius/goloxvm/internal/vmcompiler/scanner"
+	"github.com/leonardinius/goloxvm/internal/vmcompiler/tokens"
 )
 
 type Parser struct {
